# swagger.yaml
openapi: 3.0.3
info:
  title: E-Commerce App REST API
  version: 1.0.0
  description: E-Commerce App REST API using Node.js, Express, PostgresSQL. Users can perform various operations such as registering and logging into an account, browsing products, adding products to cart and viewing/placing orders. Further operations can be performed by those with admin permissions, such as retrieving details for all users, carts and orders and creating/updating/deleting products. 
...
paths:
...
  /orders:
    get:
      summary: Get all orders (ADMIN)
      description: Get all orders stored in the database
      operationId: getAllOrders
      responses:
        '200':
          description: >-
            Succesfully retrieved JSON object containing all orders for all
            users stored in the database
        '401':
          description: Authorisation error
          content:
            application/json:
              examples:
                401 Unauthorised - Not Admin:
                  value: |-
                    {
                        "message": "You are not authorised to view or edit this content!"
                    }
                401 Unauthorised - Not Logged In:
                  value: |-
                    {
                        "message": "You are not logged in. Please log in or register to continue."
                    }
        '500':
          description: Internal Server Error
          content:
            application/json:
              examples:
                500 Internal Server Error:
                  value: |-
                    {
                        "message": "connect ECONNREFUSED 127.0.0.1:5432"
                    }    
      tags:
        - Orders
  /orders/myOrders:
    get:
      summary: Get user orders
      description: Get JSON object containing all the orders for one user
      operationId: getUserOrders
      responses:
        '200':
          description: Successfully retrieved the logged in user's orders
          content:
            application/json:
              examples:
                200 OK:
                  value: |-
                    [
                        {
                            "order_id": 2,
                            "user_id": 12,
                            "email": "exampleUser3@example.com",
                            "total_price": "82.97"
                        },
                        {
                            "order_id": 3,
                            "user_id": 12,
                            "email": "exampleUser3@example.com",
                            "total_price": "85.00"
                        },
                        {
                            "order_id": 4,
                            "user_id": 12,
                            "email": "exampleUser3@example.com",
                            "total_price": "134.89"
                        }
                    ]
        '401':
          description: Authorisation error
          content:
            application/json:
              examples:
                401 Unauthorised:
                  value: |-
                    {
                        "message": "You are not logged in. Please log in or register to continue."
                    }    
        '500':
          description: Internal Server Error
          content:
            application/json:
              examples:
                500 Internal Server Error:
                  value: |-
                    {
                        "message": "connect ECONNREFUSED 127.0.0.1:5432"
                    }            
      tags:
        - Orders
  /orders/myOrders/{orderId}:
    get:
      summary: Get order by ID
      description: Filters a users list of orders down to one order by its ID
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: ID of order to get
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved details of order, specified by ID
          content:
            application/json:
              examples:
                200 OK:
                  value: |-
                    {
                        "id": 2,
                        "date": "2023-02-21T21:27:09.113Z",
                        "total_price": "82.97",
                        "user_id": 12
                    }
        '401':
          description: Authorisation error
          content:
            application/json:
              examples:
                401 Unauthorised:
                  value: |-
                    {
                        "message": "You are not logged in. Please log in or register to continue."
                    }    
        '404':
          description: No order found
          content:
            application/json:
              examples:
                404 Not Found:
                  value: |-
                    {
                        "message": "No order found with id: 5"
                    }
        '500':
          description: Internal Server Error
          content:
            application/json:
              examples:
                500 Internal Server Error:
                  value: |-
                    {
                        "message": "connect ECONNREFUSED 127.0.0.1:5432"
                    }            
      tags:
        - Orders
  /orders/myOrders/{orderId}/details:
    get:
      summary: Get products linked to order
      description: >-
        Returns a JSON object of products linked to an order, specified by order
        ID
      operationId: getOrderItemsById
      parameters:
        - name: orderId
          in: path
          description: ID of order to get details for
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved products linked to order
          content:
            application/json:
              examples:
                200 OK:
                  value: |-
                    [
                        {
                            "product_id": 5,
                            "name": "productName",
                            "description": "productDescription",
                            "category": "productCategory",
                            "quantity": 2,
                            "price_per_unit": "34.99"
                        },
                        {
                            "product_id": 6,
                            "name": "productName2",
                            "description": "productDescription2",
                            "category": "productCagetory2",
                            "quantity": 1,
                            "price_per_unit": "12.99"
                        }
                    ]
        '401':
          description: Authorisation error
          content:
            application/json:
              examples:
                401 Unauthorised:
                  value: |-
                    {
                        "message": "You are not logged in. Please log in or register to continue."
                    }    
        '404':
          description: No order found
          content:
            application/json:
              examples:
                404 Not Found:
                  value: |-
                    {
                        "message": "No order found with id: 5"
                    }
        '500':
          description: Internal Server Error
          content:
            application/json:
              examples:
                500 Internal Server Error:
                  value: |-
                    {
                        "message": "connect ECONNREFUSED 127.0.0.1:5432"
                    }              
      tags:
        - Orders
...